Package: Check for both `jqueryUi` and `jQueryUiVersion` for ThemeRoller
When constructing the ThemeRoller in the `jquery-ui.theme.css` generation
from the `lib/package.js` file, there are a number of options:
1. We have the `jQueryUiVersion` property available in `this.runtime`.
2. We have the `jqueryUi` property available in `this.runtime`.

The second happens in the jQuery UI release script. The first had to be
introduced as if one calls the `files` method of `jqueryUi`, an instance
of `JqueryUiFiles` is cached under `jqueryUi._files`, creating a circular
reference, since `JqueryUiFiles` objects keep a reference to the `JqueryUi`
ones that created them. Such a circular reference causes a crash since
`node-packager` tries to stringify `jqueryUi`.