Build: Switch from UglifyJS to SWC minify, make the minified file ES5
More recent UglifyJS versions have started converting regular functions to arrow
ones, making ES5 source file migrated to a ES2015+ minified one. We want to
avoid that even in 1.14.x as long as we keep the source file in ES5. But the
bigger problem is that we generate ES2015+ minified code even for 1.13.3
releases which support IE.

Closes gh-629
Ref mishoo/UglifyJS#5967
Ref jquery/download.jqueryui.com#629