Build: Switch from UglifyJS to SWC minify, make the minified file ES5
More recent UglifyJS versions have started converting regular functions to arrow
ones, making ES5 source file migrated to a ES2015+ minified one. We want to
avoid that even in `1.14.x` as long as we keep the source file in ES5. But the
bigger problem is that we generate ES2015+ minified code even for `1.13.3`
releases which support IE.

This change switches the minfied to SWC and ensures ES5 output. Tests
ensuring this behavior have been added.

Also, GitHub Actions were updated, as the previous versions stopped working.
To avoid this happening in the future, automatic dependabot PRs updating
actions are now enabled.

Fixes gh-629
Closes gh-630
Ref mishoo/UglifyJS#5967
Ref jquery/download.jqueryui.com#629